zebra: skip link-local routes when iterating over the RIB using the NB
The motivation for this change is that IPv6 link-local routes don't
conform to the zebra YANG module since they all have the same prefix
(fe80::/64), but zebra's YANG module require each route to have
an unique prefix (the key of the "rib" list). This violation can
cause problems when iterating over the RIB asynchronously, so skip
those routes.

At the end of the day nobody cares about link-local routes anyway :)

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>