sharpd: Add Traffic Engineering Database (TED)
Add new feature and commands to sharpd in order to collect Traffic Engineering
Database information from an IGP (OSPF or IS-IS) though the ZAPI Opaque
Message and the support of the Link State Library.

This feature serves as an example of how to code a Traffic Engineering
Database consumer and tests the mechanism.

Signed-off-by: Olivier Dugeon <olivier.dugeon@orange.com>